[42기 장주성] [MODIFY] merge 이후 naming 통일
