+++ routes/index.js
@@ -1,11 +1,11 @@
 const express = require('express');
 
-const { userRouter } = require('./users.router');
+const userRouter = require('./userRouter');
 const productRouter = require('./productRouter');
 
 const routes = express.Router();
 
-routes.use('/user', userRouter);
+routes.use('/users', userRouter.router);
 routes.use('/products', productRouter.router);
 
 module.exports = { routes };
