+++ routes/productRouter.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const productController = require('../controller/productController');
+const productController = require('../controllers/productController');
 
 router.get('/', productController.getProducts);
 router.get('/detail/:productId', productController.getProductById);
