--- routes/users.router.js
@@ -1,10 +1,0 @@
-const express = require("express");
-
-const userController = require('../controller/user-controller');
-
-const userRouter = express.Router();
-
-userRouter.post('/signup', userController.signUp);
-userRouter.post('/signin', userController.signIn);
-
-module.exports = { userRouter };
